Update Outputs
Change output submitChanges for changeCardNumber: the nomenclature was confused, it was not a submit event, but a change event

add a new Output: changeCard. It implements a return of the card data for any modified field

Included in the example a way to check the Brand of the card using the outputs.